backend/requirements.txt
pydantic-settings==2.10.1
fastapi==0.111.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
pydantic==2.7.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.3.0
databases==0.9.0
mysqlclient==2.2.1
pymysql==1.1.1
python-dotenv==1.0.0
